requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118
numpy
datasets==4.3.0
diffusers==0.32.2
tqdm==4.67.1
pika==1.3.2
python3-pika==0.9.14
scipy==1.14.1
sentence-transformers==5.1.0
python-dotenv==1.1.0
matplotlib==3.9.2
kagglehub==0.3.11
pillow==10.2.0
